Fixed CMakeLists
