+++ libs/yaffs_glue/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(NAME yaffs_glue)
 
 add_library(${NAME} STATIC    
-    yaffs_os_glue.c
+    yaffs_os_glue.cpp
 )
 
 target_link_libraries(${NAME} PUBLIC
